platformer demo is buildable

+++ Editor/Platform/BardicBuildPreprocessor.cs
@@ -14,13 +14,15 @@ namespace BardicBytes.BardicFrameworkEditor.Platform
             var p = System.IO.Path.Combine(Application.dataPath, "Data","buildinfo.json");
             //Debug.Log("loading "+p);
             //TextAsset ta = AssetDatabase.LoadAssetAtPath<TextAsset>("Assets/Data/buildinfo.txt");
-            string json = System.IO.File.ReadAllText(p);
-            //Debug.Log("loaded "+ta);
-            var bi = JsonUtility.FromJson<BuildInfo>(json);
-            bi.Update();
-            System.IO.File.WriteAllText(System.IO.Path.Combine(Application.dataPath, p), bi.ToJson());
-
-            Debug.Log(bi.ToString()+"\nBardicBuildPreprocessor.OnPreprocessBuild for target " + report.summary.platform + " at path " + report.summary.outputPath);
+            if (System.IO.File.Exists(p))
+            {
+                string json = System.IO.File.ReadAllText(p);
+                //Debug.Log("loaded "+ta);
+                var bi = JsonUtility.FromJson<BuildInfo>(json);
+                bi.Update();
+                System.IO.File.WriteAllText(System.IO.Path.Combine(Application.dataPath, p), bi.ToJson());
+                Debug.Log(bi.ToString()+"\nBardicBuildPreprocessor.OnPreprocessBuild for target " + report.summary.platform + " at path " + report.summary.outputPath);
+            }
         }
     }
 }
